Ajustes

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,7 +20,7 @@ class CreateUsersTable extends Migration
             $table->string('name');
             $table->string('email')->unique();
             $table->integer('cedula')->nullable();
-            $table->integer('telefono')->nullable();
+            $table->bigInteger('telefono')->nullable();
             $table->string('direccion')->nullable();
             $table->integer('NIT')->nullable();
             $table->timestamp('email_verified_at')->nullable();
